pyproject.toml
[project]
name = "rat"
version = "0.1.0"
description = "A powerful AI reasoning chain CLI tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "groq>=0.3.0",
    "openai>=1.59.9",
    "prompt-toolkit>=3.0.50",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[project.scripts]
rat = "rat.rat:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rat"]
